[MOBI-33-landing-page-bottom-mobile-publishing] Refactor: 겹치는 타입 한 곳에서 관리
