+++ apps/academy/src/widget/landing/constants/index.ts
@@ -1,5 +1,6 @@
 export {
   ADVANCED_COURSES,
+  type ContentCardProps,
   EMPLOYMENT_SUPPORT_COURSES,
   INITIAL_COURSES,
   PROJECT_COURSES
